A big refactor to elimitate view decorators
 * View decorators have been either inlined or moved to a service
 * The new services is LTIHypothesisBridge and is responsible for
   updating H with LTI related groups and users
 * This meant removing a lot of decorator wrapper related stuff
 * I've split the single BasicLTIView class into three:
   * LTILaunchViewBaseClass
   * ConfigureModuleItemView
   * BasicLTILaunchViews
 * I've split out the tests for basic_lti_launch as they were huge
